track(publisher): send keyframe requests on change
